Use eCFR for FAR references (#16)
This modifies the URI scheme for directing users to FAR source material.
Previously users were sent to the www.law.cornell.edu/cfr site, which was
(vastly) easier to do thanks to its simple and consistent URI structure.
However, it is not the authoritative source--that would be
www.ecfr.gov/current/title-14/. Following this commit users will now be directed
to the latter.

Unfortunately, the URI scheme for the eCFR website is *muuuuch* more
complicated. While the Cornell site is able to properly direct requests using
only the part and section numbers, the eCFR site requires the full
specification: the chapter, subchapter, part, subpart, subject group (for
part 91) and section. I wanted the API for the `<FAR>` component to remain
simple (i.e. specifying part and section should be sufficient), and this
necessitated looking up the rest of the section identifiers in the background.

To support this, I created the `build_far_json.py` script, which generates the
`far.json` data file. The data file is included in the bundle and referred to in
two places: the `references.ts` file where the convoluted URI is generated, and
within the `<FAR>` component where the on-page textual representation is built
(including tooltip text to show the cited regulation's label). A somewhat
elaborate type namespace was also created to describe the `far.json` object.